requirements.txt
Pillow>=10.0.0
google-genai>=1.0.0
dotenv
pygame>=2.0.0
elevenlabs